k2_quant_web/backend/requirements.txt
# K2 Quant Web Backend Requirements
# Python 3.10+

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
alembic>=1.12.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b0
TA-Lib>=0.4.28
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# API and Networking
requests>=2.31.0
aiohttp>=3.8.0
websockets>=11.0
httpx>=0.25.0

# AI/ML Integration
openai>=1.0.0
anthropic>=0.7.0
tiktoken>=0.5.0

# ML Libraries
yfinance>=0.2.0
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Background Tasks
celery>=5.3.0
redis>=5.0.0
flower>=2.0.0

# Utilities
pytz>=2023.3
python-dateutil>=2.8.2
schedule>=1.2.0

# CORS and Security
python-cors>=1.0.0
slowapi>=0.1.9

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
black>=23.3.0
flake8>=6.0.0
mypy>=1.5.0